requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-documentai
google-cloud-storage
python-dotenv